requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.8
backoff==2.2.1
bandit==1.9.4
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.5
distro==1.9.0
docutils==0.22
dparse==0.6.4
durationpy==0.10
esprima==4.0.1
fastapi==0.116.1
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.28.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grokipedia-api==0.3.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
jh2==5.0.9
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.3
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
markdown2==2.5.4
MarkupSafe==3.0.2
marshmallow==4.2.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.3
nicegui==2.22.0
niquests==3.14.1
nltk==3.9.3
numpy==2.3.2
oauthlib==3.3.1
ollama==0.5.1
onnxruntime==1.22.1
openai==1.98.0
openmeteo_requests==1.6.0
openmeteo_sdk==1.20.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.3.1
patch-ng==1.18.1
platformdirs==4.3.8
posthog==5.4.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pscript==0.7.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==3.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymdown-extensions==10.16.1
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
qh3==1.5.3
referencing==0.36.2
regex==2026.2.19
requests==2.32.4
requests-cache==1.2.1
requests-oauthlib==2.0.0
retry-requests==2.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.19.1
ruff==0.15.3
safety==3.7.0
safety-schemas==0.0.16
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
stevedore==5.7.0
style==1.1.0
sympy==1.14.0
tenacity==8.5.0
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.14.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
update==0.0.1
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.5.0
urllib3-future==2.13.900
uvicorn==0.35.0
vbuild==0.8.2
waitress==3.0.2
wassima==1.2.2
watchfiles==1.1.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
